requirements.txt
# VideoDetective Dependencies
# ========================

# Core ML/DL
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
torch>=1.9.0
transformers>=4.30.0

# Image/Video Processing
pillow>=9.0.0
decord>=0.6.0
matplotlib>=3.5.0

# LLM Client
openai>=1.0.0

# Configuration
python-dotenv>=0.19.0

# Data Processing
pandas>=1.3.0

# ASR (Automatic Speech Recognition)
openai-whisper>=20231117

# System requirements (not pip installable):
# - ffmpeg: Required for audio extraction. Install via:
#   - Ubuntu/Debian: sudo apt install ffmpeg
#   - macOS: brew install ffmpeg
#   - Windows: Download from https://ffmpeg.org/download.html

# Optional: OCR support (for text detection in videos)
# easyocr>=1.6.0
